requirements.txt
paho-mqtt==1.5.1
pigpio-dht==0.3.6
psycopg2==2.8.6
